Allow command() arguments to be created via a function call.
Allow a call to command() to have arguments defined by the
result of a function.  So this is now possible:

command(() => ["add", filename])

Where filename can be defined in a previous .do() statement.

This allows tests to build command arguments dynamically rather
that at the time the test is compiled.