feat: add a `returned` value to the test context on commands
While using the `command` test helper, it can be nice to check the
returned value of the executed command.

CLI commands don't usually return any specific values, but if you
consider reusing commands (via imports) to compose small commands to
build bigger commands it can be nice to be able to check the output of
a command in a test context.

I'm aware calling commands programmatically is not recommended by the
oclif documentation, however it's still a documented feature here:
https://oclif.io/docs/running_programmatically#calling-commands-directly